add eval and ips (#4947)
* del unused code

* add eval

* add resume

* fix error
