+++ ppstructure/vqa/requirements.txt
@@ -1,2 +1,3 @@
 sentencepiece
 yacs
+seqeval
